Fixes #18951 - Assigning Ansible roles over the API
At the moment, it is only possible to assign Ansible roles from the
Foreman UI. Due to this reason, Foreman cannot be used as a pure backend
service and it cannot be configured externally. In this feature, new
endpoints are added to the API, which allow setting roles as well as
listing assigned roles.
